use the trilinos supplied FindZoltan cmake module
while it may not be perfect, at least it is the way to detect zoltan
that is officially supported by the trilinos project.

In order to make this work, the target for the "zoltan" test needed to
be renamed to avoid a clash of target names with the official
FindZoltan module.