CpGrid: sort the number of cells per process by the total number of cells on a given rank
this was sorted by rank. IMHO it makes much more sense to sort this by
the number of cells on each rank because this allows to quickly
calculate the "quality" of the partition, e.g., the number of cells on
the rank with most of them divided by the number of cells on the rank
with the fewest cells.